backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
supabase==2.0.2

# HTTP Clients
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0
lxml==4.9.3

# AI/ML
openai==1.3.7
google-generativeai==0.3.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Email
google-api-python-client==2.110.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
msal==1.25.0

# Utilities
python-dotenv==1.0.0
python-slugify==8.0.1
pandas==2.1.4
numpy==1.24.4
Pillow==10.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
structlog==23.2.0